Installed MongoDb on host and fixed a crash in a converter

+++ src/converter/convert/to_mp3.py
@@ -16,26 +16,29 @@ def start(message, fs_videos, fs_mp3s, channel):
     tf.close()
 
     # write audio to a file
-    tf_path = tempfile.gettempdir() + f"/{message['video_fid']}.mp3"
-    audio.write_audiofile(tf_path)
+    if audio != None:
+        tf_path = tempfile.gettempdir() + f"/{message['video_fid']}.mp3"
+        audio.write_audiofile(tf_path)
+    
+        # save file to mongo
+        f = open(tf_path, "rb")
+        data = f.read()
+        fid = fs_mp3s.put(data)
+        f.close()
+        os.remove(tf_path)
 
-    # save file to mongo
-    f = open(tf_path, "rb")
-    data = f.read()
-    fid = fs_mp3s.put(data)
-    f.close()
-    os.remove(tf_path)
-
-    message["mp3_id"] = str(fid)
-    try:
-        channel.basic_publish(
-            exchange="",
-            routing_key=os.environ.get("MP3_QUEUE"),
-            body=json.dumps(message),
-            properties=pika.BasicProperties(
-                delivery_mode=pika.spec.PERSISTENT_DELIVERY_MODE
-            ),
-        )
-    except Exception as err:
-        fs_mp3s.delete(fid)
-        return "failed to publish message"
+        message["mp3_id"] = str(fid)
+        try:
+            channel.basic_publish(
+                exchange="",
+                routing_key=os.environ.get("MP3_QUEUE"),
+                body=json.dumps(message),
+                properties=pika.BasicProperties(
+                    delivery_mode=pika.spec.PERSISTENT_DELIVERY_MODE
+                ),
+            )
+        except Exception as err:
+            fs_mp3s.delete(fid)
+            return "failed to publish message"
+    else:
+        return "Sorry, this video contains no audio."
